リファクタリング途中
