+++ resources/views/subs/flash-message-danger.blade.php
@@ -1,0 +1,8 @@
+@if (session('flash_message'))
+  <div class="alert alert-danger alert-dismissible fade show" role="alert">
+    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+      <span aria-hidden="true">&times;</span>
+    </button>
+    {{ session('flash_message') }}
+  </div>
+@endif
