+++ resources/views/user/edit.blade.php
@@ -4,14 +4,7 @@
   <div class="col mx-3">
     <div id="crop-avatar">
 
-      @if (session('flash_message'))
-        <div class="alert alert-success alert-dismissible fade show" role="alert">
-          <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-            <span aria-hidden="true">&times;</span>
-          </button>
-          {{ session('flash_message') }}
-        </div>
-      @endif
+      @include('subs.flash-message-success')
 
       <h4>プロフィールを編集</h4>
